Nectron/requirements.txt
dearpygui==1.11.1
google-generativeai==0.8.1
lark==1.1.9
lxml==5.3.0
openai==1.96.1
pandas==2.2.2
PyYAML==6.0.2